+++ Modules/Intent.Modules.Application.FluentValidation.Dtos/Templates/ValidationServiceInterface/ValidationServiceInterfaceTemplateRegistration.cs
@@ -40,6 +40,7 @@ namespace Intent.Modules.Application.FluentValidation.Dtos.Templates.ValidationS
             base.DoRegistration(registry, application);
         }
 
+        [IntentManaged(Mode.Fully)]
         public override ITemplate CreateTemplateInstance(IOutputTarget outputTarget)
         {
             return new ValidationServiceInterfaceTemplate(outputTarget);