+++ Modules/Intent.Modules.Application.FluentValidation.Dtos/Templates/ValidatorProviderInterface/ValidatorProviderInterfaceTemplatePartial.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using Intent.Engine;
@@ -29,7 +30,7 @@ namespace Intent.Modules.Application.FluentValidation.Dtos.Templates.ValidatorPr
                     inter.AddMethod($"IValidator<T>", "GetValidator", method => method.AddGenericParameter("T"));
                 });
         }
-        
+
         // public override bool CanRunTemplate()
         // {
         //     var validatorTemplates = ExecutionContext.FindTemplateInstances<ICSharpFileBuilderTemplate>(TemplateDependency.OnTemplate("Application.Validation.Dto"))