+++ Modules/Intent.Modules.Application.FluentValidation/Api/DTOFieldModelStereotypeExtensions.cs
@@ -15,14 +15,14 @@ namespace Intent.Application.FluentValidation.Api
     {
         public static Validations GetValidations(this DTOFieldModel model)
         {
-            var stereotype = model.GetStereotype("Validations");
+            var stereotype = model.GetStereotype("4b54a612-2664-4493-a1f7-dc0623aa03da");
             return stereotype != null ? new Validations(stereotype) : null;
         }
 
 
         public static bool HasValidations(this DTOFieldModel model)
         {
-            return model.HasStereotype("Validations");
+            return model.HasStereotype("4b54a612-2664-4493-a1f7-dc0623aa03da");
         }
 
         public static bool TryGetValidations(this DTOFieldModel model, out Validations stereotype)
@@ -33,7 +33,7 @@ namespace Intent.Application.FluentValidation.Api
                 return false;
             }
 
-            stereotype = new Validations(model.GetStereotype("Validations"));
+            stereotype = new Validations(model.GetStereotype("4b54a612-2664-4493-a1f7-dc0623aa03da"));
             return true;
         }
 
@@ -82,6 +82,11 @@ namespace Intent.Application.FluentValidation.Api
             public string Max()
             {
                 return _stereotype.GetProperty<string>("Max");
+            }
+
+            public string RegularExpression()
+            {
+                return _stereotype.GetProperty<string>("Regular Expression");
             }
 
             public string Predicate()