+++ Modules/Intent.Modules.Application.MediatR.CRUD.Tests/Templates/TemplateExtensions.cs
@@ -1,4 +1,6 @@
 using System.Collections.Generic;
+using Intent.Modelers.Domain.Api;
+using Intent.Modelers.Services.CQRS.Api;
 using Intent.Modules.Application.MediatR.CRUD.Tests.Templates.Assertions.AssertionClass;
 using Intent.Modules.Application.MediatR.CRUD.Tests.Templates.Extensions.RepositoryExtensions;
 using Intent.Modules.Application.MediatR.CRUD.Tests.Templates.FluentValidation.FluentValidationTest;
@@ -24,12 +26,12 @@ namespace Intent.Modules.Application.MediatR.CRUD.Tests.Templates
 {
     public static class TemplateExtensions
     {
-        public static string GetAssertionClassName<T>(this IIntentTemplate<T> template) where T : Intent.Modelers.Domain.Api.ClassModel
+        public static string GetAssertionClassName<T>(this IIntentTemplate<T> template) where T : ClassModel
         {
             return template.GetTypeName(AssertionClassTemplate.TemplateId, template.Model);
         }
 
-        public static string GetAssertionClassName(this IIntentTemplate template, Intent.Modelers.Domain.Api.ClassModel model)
+        public static string GetAssertionClassName(this IIntentTemplate template, ClassModel model)
         {
             return template.GetTypeName(AssertionClassTemplate.TemplateId, model);
         }
@@ -39,122 +41,122 @@ namespace Intent.Modules.Application.MediatR.CRUD.Tests.Templates
             return template.GetTypeName(RepositoryExtensionsTemplate.TemplateId);
         }
 
-        public static string GetFluentValidationTestName<T>(this IIntentTemplate<T> template) where T : Intent.Modelers.Services.CQRS.Api.CommandModel
+        public static string GetFluentValidationTestName<T>(this IIntentTemplate<T> template) where T : CommandModel
         {
             return template.GetTypeName(FluentValidationTestTemplate.TemplateId, template.Model);
         }
 
-        public static string GetFluentValidationTestName(this IIntentTemplate template, Intent.Modelers.Services.CQRS.Api.CommandModel model)
+        public static string GetFluentValidationTestName(this IIntentTemplate template, CommandModel model)
         {
             return template.GetTypeName(FluentValidationTestTemplate.TemplateId, model);
         }
 
-        public static string GetNestedCreateCommandHandlerTestsName<T>(this IIntentTemplate<T> template) where T : Intent.Modelers.Services.CQRS.Api.CommandModel
+        public static string GetNestedCreateCommandHandlerTestsName<T>(this IIntentTemplate<T> template) where T : CommandModel
         {
             return template.GetTypeName(NestedCreateCommandHandlerTestsTemplate.TemplateId, template.Model);
         }
 
-        public static string GetNestedCreateCommandHandlerTestsName(this IIntentTemplate template, Intent.Modelers.Services.CQRS.Api.CommandModel model)
+        public static string GetNestedCreateCommandHandlerTestsName(this IIntentTemplate template, CommandModel model)
         {
             return template.GetTypeName(NestedCreateCommandHandlerTestsTemplate.TemplateId, model);
         }
 
-        public static string GetNestedDeleteCommandHandlerTestsName<T>(this IIntentTemplate<T> template) where T : Intent.Modelers.Services.CQRS.Api.CommandModel
+        public static string GetNestedDeleteCommandHandlerTestsName<T>(this IIntentTemplate<T> template) where T : CommandModel
         {
             return template.GetTypeName(NestedDeleteCommandHandlerTestsTemplate.TemplateId, template.Model);
         }
 
-        public static string GetNestedDeleteCommandHandlerTestsName(this IIntentTemplate template, Intent.Modelers.Services.CQRS.Api.CommandModel model)
+        public static string GetNestedDeleteCommandHandlerTestsName(this IIntentTemplate template, CommandModel model)
         {
             return template.GetTypeName(NestedDeleteCommandHandlerTestsTemplate.TemplateId, model);
         }
 
-        public static string GetNestedGetAllQueryHandlerTestsName<T>(this IIntentTemplate<T> template) where T : Intent.Modelers.Services.CQRS.Api.QueryModel
+        public static string GetNestedGetAllQueryHandlerTestsName<T>(this IIntentTemplate<T> template) where T : QueryModel
         {
             return template.GetTypeName(NestedGetAllQueryHandlerTestsTemplate.TemplateId, template.Model);
         }
 
-        public static string GetNestedGetAllQueryHandlerTestsName(this IIntentTemplate template, Intent.Modelers.Services.CQRS.Api.QueryModel model)
+        public static string GetNestedGetAllQueryHandlerTestsName(this IIntentTemplate template, QueryModel model)
         {
             return template.GetTypeName(NestedGetAllQueryHandlerTestsTemplate.TemplateId, model);
         }
 
-        public static string GetNestedGetByIdQueryHandlerTestsName<T>(this IIntentTemplate<T> template) where T : Intent.Modelers.Services.CQRS.Api.QueryModel
+        public static string GetNestedGetByIdQueryHandlerTestsName<T>(this IIntentTemplate<T> template) where T : QueryModel
         {
             return template.GetTypeName(NestedGetByIdQueryHandlerTestsTemplate.TemplateId, template.Model);
         }
 
-        public static string GetNestedGetByIdQueryHandlerTestsName(this IIntentTemplate template, Intent.Modelers.Services.CQRS.Api.QueryModel model)
+        public static string GetNestedGetByIdQueryHandlerTestsName(this IIntentTemplate template, QueryModel model)
         {
             return template.GetTypeName(NestedGetByIdQueryHandlerTestsTemplate.TemplateId, model);
         }
 
-        public static string GetNestedUpdateCommandHandlerTestsName<T>(this IIntentTemplate<T> template) where T : Intent.Modelers.Services.CQRS.Api.CommandModel
+        public static string GetNestedUpdateCommandHandlerTestsName<T>(this IIntentTemplate<T> template) where T : CommandModel
         {
             return template.GetTypeName(NestedUpdateCommandHandlerTestsTemplate.TemplateId, template.Model);
         }
 
-        public static string GetNestedUpdateCommandHandlerTestsName(this IIntentTemplate template, Intent.Modelers.Services.CQRS.Api.CommandModel model)
+        public static string GetNestedUpdateCommandHandlerTestsName(this IIntentTemplate template, CommandModel model)
         {
             return template.GetTypeName(NestedUpdateCommandHandlerTestsTemplate.TemplateId, model);
         }
 
-        public static string GetCreateCommandHandlerTestsName<T>(this IIntentTemplate<T> template) where T : Intent.Modelers.Services.CQRS.Api.CommandModel
+        public static string GetCreateCommandHandlerTestsName<T>(this IIntentTemplate<T> template) where T : CommandModel
         {
             return template.GetTypeName(CreateCommandHandlerTestsTemplate.TemplateId, template.Model);
         }
 
-        public static string GetCreateCommandHandlerTestsName(this IIntentTemplate template, Intent.Modelers.Services.CQRS.Api.CommandModel model)
+        public static string GetCreateCommandHandlerTestsName(this IIntentTemplate template, CommandModel model)
         {
             return template.GetTypeName(CreateCommandHandlerTestsTemplate.TemplateId, model);
         }
 
-        public static string GetDeleteCommandHandlerTestsName<T>(this IIntentTemplate<T> template) where T : Intent.Modelers.Services.CQRS.Api.CommandModel
+        public static string GetDeleteCommandHandlerTestsName<T>(this IIntentTemplate<T> template) where T : CommandModel
         {
             return template.GetTypeName(DeleteCommandHandlerTestsTemplate.TemplateId, template.Model);
         }
 
-        public static string GetDeleteCommandHandlerTestsName(this IIntentTemplate template, Intent.Modelers.Services.CQRS.Api.CommandModel model)
+        public static string GetDeleteCommandHandlerTestsName(this IIntentTemplate template, CommandModel model)
         {
             return template.GetTypeName(DeleteCommandHandlerTestsTemplate.TemplateId, model);
         }
 
-        public static string GetGetAllPaginationQueryHandlerTestsName<T>(this IIntentTemplate<T> template) where T : Intent.Modelers.Services.CQRS.Api.QueryModel
+        public static string GetGetAllPaginationQueryHandlerTestsName<T>(this IIntentTemplate<T> template) where T : QueryModel
         {
             return template.GetTypeName(GetAllPaginationQueryHandlerTestsTemplate.TemplateId, template.Model);
         }
 
-        public static string GetGetAllPaginationQueryHandlerTestsName(this IIntentTemplate template, Intent.Modelers.Services.CQRS.Api.QueryModel model)
+        public static string GetGetAllPaginationQueryHandlerTestsName(this IIntentTemplate template, QueryModel model)
         {
             return template.GetTypeName(GetAllPaginationQueryHandlerTestsTemplate.TemplateId, model);
         }
 
-        public static string GetGetAllQueryHandlerTestsName<T>(this IIntentTemplate<T> template) where T : Intent.Modelers.Services.CQRS.Api.QueryModel
+        public static string GetGetAllQueryHandlerTestsName<T>(this IIntentTemplate<T> template) where T : QueryModel
         {
             return template.GetTypeName(GetAllQueryHandlerTestsTemplate.TemplateId, template.Model);
         }
 
-        public static string GetGetAllQueryHandlerTestsName(this IIntentTemplate template, Intent.Modelers.Services.CQRS.Api.QueryModel model)
+        public static string GetGetAllQueryHandlerTestsName(this IIntentTemplate template, QueryModel model)
         {
             return template.GetTypeName(GetAllQueryHandlerTestsTemplate.TemplateId, model);
         }
 
-        public static string GetGetByIdQueryHandlerTestsName<T>(this IIntentTemplate<T> template) where T : Intent.Modelers.Services.CQRS.Api.QueryModel
+        public static string GetGetByIdQueryHandlerTestsName<T>(this IIntentTemplate<T> template) where T : QueryModel
         {
             return template.GetTypeName(GetByIdQueryHandlerTestsTemplate.TemplateId, template.Model);
         }
 
-        public static string GetGetByIdQueryHandlerTestsName(this IIntentTemplate template, Intent.Modelers.Services.CQRS.Api.QueryModel model)
+        public static string GetGetByIdQueryHandlerTestsName(this IIntentTemplate template, QueryModel model)
         {
             return template.GetTypeName(GetByIdQueryHandlerTestsTemplate.TemplateId, model);
         }
 
-        public static string GetUpdateCommandHandlerTestsName<T>(this IIntentTemplate<T> template) where T : Intent.Modelers.Services.CQRS.Api.CommandModel
+        public static string GetUpdateCommandHandlerTestsName<T>(this IIntentTemplate<T> template) where T : CommandModel
         {
             return template.GetTypeName(UpdateCommandHandlerTestsTemplate.TemplateId, template.Model);
         }
 
-        public static string GetUpdateCommandHandlerTestsName(this IIntentTemplate template, Intent.Modelers.Services.CQRS.Api.CommandModel model)
+        public static string GetUpdateCommandHandlerTestsName(this IIntentTemplate template, CommandModel model)
         {
             return template.GetTypeName(UpdateCommandHandlerTestsTemplate.TemplateId, model);
         }